+++ Pythia.Core.Plugin/Analysis/LenSupplierTokenFilter.cs
@@ -3,6 +3,7 @@ using Fusi.Tools.Config;
 using Pythia.Core.Analysis;
 using System;
 using System.Globalization;
+using Fusi.Tools;
 
 namespace Pythia.Core.Plugin.Analysis
 {
@@ -27,8 +28,9 @@ namespace Pythia.Core.Plugin.Analysis
         /// <param name="position">The position which will be assigned to
         /// the resulting token, provided that it's not empty. Not used.
         /// </param>
+        /// <param name="context">The optional context. Not used.</param>
         /// <exception cref="ArgumentNullException">token</exception>
-        public void Apply(Token token, int position)
+        public void Apply(Token token, int position, IHasDataDictionary? context = null)
         {
             if (token == null) throw new ArgumentNullException(nameof(token));
             if (string.IsNullOrEmpty(token.Value)) return;