+++ Pythia.Core.Plugin/Analysis/WhitespaceTokenizer.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text;
+using Fusi.Tools;
 using Fusi.Tools.Config;
 using Pythia.Core.Analysis;