+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- 2022-12-08:
+  - added context to filters.
+  - added UDP project.
 - 2022-11-10: upgraded to NET 7.
 - 2022-11-05:
   - refactored all projects for nullability check.